refactor: 포스터 이미지 맵 패키지이동
